refactor: 리뷰 내용 반영 및 코드 구조 개선
