--- src/main/java/com/ninedocs/userserver/application/dockercomposetest/dto/TestResponse.java
@@ -1,12 +1,0 @@
-package com.ninedocs.userserver.application.dockercomposetest.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@Data
-public class TestResponse {
-
-  private String email;
-}
