--- src/main/java/com/ninedocs/userserver/application/emailverificationcode/dto/EmailVerificationCodeResponse.java
@@ -1,10 +1,0 @@
-package com.ninedocs.userserver.application.emailverificationcode.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Data;
-
-@Data
-public class EmailVerificationCodeResponse {
-
-  private String verificationCodeExpiredAt;
-}
