+++ src/main/java/com/ninedocs/userserver/common/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.ninedocs.userserver.common.exception;
 
-import com.ninedocs.userserver.common.response.ApiResponse;
+import com.ninedocs.userserver.common.presentation.dto.ApiResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -10,7 +10,7 @@ public class GlobalExceptionHandler {
 
   @ExceptionHandler(CustomException.class)
   public ResponseEntity<ApiResponse<Void>> handleCustomException(CustomException e) {
-    return ResponseEntity.badRequest().body(
+    return ResponseEntity.ok().body(
         ApiResponse.error(e.getErrorCode())
     );
   }
