+++ src/main/java/com/ninedocs/userserver/common/util/DateTimeUtil.java
@@ -1,0 +1,12 @@
+package com.ninedocs.userserver.common.util;
+
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+
+public class DateTimeUtil {
+
+  public static LocalDateTime convertLongToLocalDateTime(long timestamp) {
+    return LocalDateTime.ofInstant(Instant.ofEpochMilli(timestamp), ZoneId.systemDefault());
+  }
+}
