--- src/main/java/com/ninedocs/userserver/persistence/UserRepository.java
@@ -1,10 +1,0 @@
-package com.ninedocs.userserver.persistence;
-
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-  Optional<User> findByEmail(String email);
-}
