--- src/main/java/com/ninedocs/userserver/presentation/HealthController.java
@@ -1,13 +1,0 @@
-package com.ninedocs.userserver.presentation;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HealthController {
-  
-  @GetMapping("/api/v1/system/health")
-  public String health() {
-    return "Healthy";
-  }
-}
