+++ src/main/java/com/ninedocs/userserver/user/application/dockercomposetest/dto/TestResponse.java
@@ -1,0 +1,12 @@
+package com.ninedocs.userserver.user.application.dockercomposetest.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@Data
+public class TestResponse {
+
+  private String email;
+}
