+++ src/main/java/com/ninedocs/userserver/user/application/emailverificationcode/EmailDuplicationChecker.java
@@ -1,0 +1,16 @@
+package com.ninedocs.userserver.user.application.emailverificationcode;
+
+import com.ninedocs.userserver.user.persistence.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class EmailDuplicationChecker {
+
+  private final UserRepository userRepository;
+
+  public Boolean isDuplicateEmail(String email) {
+    return userRepository.findByEmail(email).isPresent();
+  }
+}
