+++ src/main/java/com/ninedocs/userserver/user/application/emailverificationcode/EmailNotificationService.java
@@ -1,0 +1,8 @@
+package com.ninedocs.userserver.user.application.emailverificationcode;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class EmailNotificationService {
+
+}
