+++ src/main/java/com/ninedocs/userserver/user/persistence/UserRepository.java
@@ -1,0 +1,10 @@
+package com.ninedocs.userserver.user.persistence;
+
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+  Optional<User> findByEmail(String email);
+}
