+++ src/main/java/com/ninedocs/userserver/user/presentation/HealthController.java
@@ -1,0 +1,13 @@
+package com.ninedocs.userserver.user.presentation;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class HealthController {
+
+  @GetMapping("/api/v1/system/health")
+  public String health() {
+    return "Healthy";
+  }
+}
